+++ src/gui.cpp
@@ -25,7 +25,7 @@
 #include "imgui.h"
 #include "utl_log.h"
 #include "scr_lua_console.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "ezy_io.h"
 #include "scr_python_console.h"
 #include "shp_info.h"
