+++ src/gui_mode.cpp
@@ -16,7 +16,7 @@
 #include "imgui.h"
 #include "imgui_internal.h"
 #include "mode.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch.h"
 #include "utl_occt.h"
