+++ src/shp_polar_dup.cpp
@@ -10,7 +10,7 @@
 #include "utl_geom.h"
 #include "gui.h"
 #include "mode.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch.h"
 #include "sketch_nodes.h"
