+++ src/sketch_delta.cpp
@@ -8,7 +8,7 @@
 
 #include "utl_dbg.h"
 #include "utl_geom.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch.h"
 #include "sketch_nodes.h"
 #include "sketch_edge.h"
