+++ src/sketch_dims.cpp
@@ -10,7 +10,7 @@
 
 #include "gui.h"
 #include "mode.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch.h"
 #include "sketch_delta.h"
 #include "sketch_edge.h"
