+++ src/sketch_json.cpp
@@ -18,7 +18,7 @@
 #include "sketch_nodes.h"
 #include "sketch_underlay.h"
 #include "ezy_asset_store.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "utl_json.h"
 
 using json = nlohmann::json;
