+++ src/sketch_node_marks.cpp
@@ -5,7 +5,7 @@
 
 #include "gui.h"
 #include "mode.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch.h"
 #include "sketch_nodes.h"
 #include "sketch_topo.h"
