+++ src/sketch_nodes.cpp
@@ -13,7 +13,7 @@
 #include "utl_dbg.h"
 #include "utl_geom.h"
 #include "imgui.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 
 namespace
 {
