+++ src/sketch_tools.cpp
@@ -12,7 +12,7 @@
 
 #include "gui.h"
 #include "mode.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch.h"
 #include "sketch_delta.h"
 #include "sketch_edge.h"
