+++ src/sketch_topo.cpp
@@ -19,7 +19,7 @@
 
 #include <glm/glm.hpp>
 
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch.h"
 #include "sketch_delta.h"
 #include "sketch_edge.h"
