+++ src/sketch_underlay.cpp
@@ -24,7 +24,7 @@
 
 #include "utl_geom.h"
 #include "ezy_asset_store.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch_nodes.h"
 #include "utl.h"
