+++ tests/sketch_tests.cpp
@@ -13,7 +13,7 @@
 
 #include "utl_geom.h"
 #include "gui.h"
-#include "occt_view.h"
+#include "gui_occt_view.h"
 #include "sketch.h"
 #include "sketch_delta.h"
 #include "sketch_json.h"
